Refactor config setup and logger handling
Replaces the Setup class with ConfigInit for configuration initialization, updates imports and references accordingly, and simplifies logger injection by accepting any logger-like object. Removes unused and commented-out code, and deletes setup.py.
